Create cycling
In brief, this function performs a pairwise comparison and calculates the number of mutations of all sequences from a given library against their consensus.

The function is highly customized to handle with Cycling project (Vierna et al. 2016) files (i.e. names and reads size). It uses ape  functions that requires sequences to be identical in size. So, sequences with insertions or deletions are first aligned against consensus prior calculations. Insertions and deletions are treated as mutations.